[UPDATE & ADD] - updated prototypes of instructions functions (must make it all) and add the tab of pointer on function

+++ src/instructions/st.c
@@ -7,7 +7,7 @@
 
 #include "my.h"
 
-int execute_st(corewar_t *c, uint8_t *args, int nbr_player)
+int execute_st(corewar_t *c, uint8_t *args, size_t nbr_player)
 {
     if (!c || !args)
         return ERROR;
